Adjust dimension of blank pages
With PDF readers (such as zathura and Sumatra PDF), when you fit the
display to page size, it does not fit properly if the document
contains pages that are different in size from other pages.

Also, when using the double-page spread view, if the left page and the
right page are not alternated, the left and right sides are switched
from the middle, making it difficult to read.

Therefore, we matched the size of blank pages that were too large with
other pages and inserted missing blank pages. We used GNU make to fix
the PDF file, since the PDF is about GNU make.

The Makefile is available at:

https://gist.github.com/yoheiy/51aff1c280e3ac5aa9b626f9d7e2442e